+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Implement `xsd:boolean` data type codec
 - Implement `xsd:string` data type codec
 - Implement `xsd:nmtoken` data type codec
+- Implement `xsd:normalizedString` data type codec